Spaceship & Laser Designs
Created Illustrator Files for the Spaceship and the Laser designs of
both the Aliens and the Spaceship. These were then exported as PNGs so
as to have transparent backgrounds for Unity and then added to the
appropriate Prefabs. The Prefabs were given appropriate scales as well
as the Textures were assigned as Unlit>Transparent so as to show
appropriately for the game.