Alien, Boss & Boosts Designs
Created and Implemented Illustrator files and exported as PNGs so as to
use them in the game. The alien and the boss have been uploaded as
sprite sheets so as to have the 4 frame animation like the spaceship.
They've been imported as Unlit/Transparent textures as well as arrange
for HD purposes.