Help Menu & Testing Highscore
After various experiments in how to display 2D Textures in Unity, I
managed to have an effective and illustrative Help Menu, showing the
user which is the Spaceship and Aliens and what buttons they need to
press to play the game. I've also started experimenting and researching
how the IO system works of Unity so as to keep a Highscore file and
display it accordingly.