Sounds, Help Menu, Space Invaders Logo
The sounds were recorded appropriately by using an online virtual
synthesizer from www.audiosauna.com/studio. With a mic attached to a PC,
I used the Laser and Spaceship presets from the synthesizer so as to
record the required sounds. I experimented with various settings such as
stretch, pitch, decay and more so as to create the sounds I desire.
After recording them, I imported them in Adobe Audition CS6 so as to
edit them one by one and add reverbs, echoes and remove any background
noise.

The Help Menu was implemented so as the user would know which buttons to
press, who is the player and who is the enemy. I did this in a simple
but illustrative manner for a wider range of audience.

The Space Invaders Logo was also added to the beginning of the game in
the Main Menu by using the imported font Starcraft.